+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0] - 2025-01-01
+### Added
+- Browser version of the renderer and compiled templates.  Distribution files are `dist/govcyCompiledTemplates.browser.js`, `dist#govcyFrontendRenderer.browser.js`
+- Test page for browser version on `client\testclientjson.html`
+- Test script that checks browser it renders on `test\moca\render.client.test.mjs` 
+
+### Changed
+- Rollup build script to create the browser version
+- Test scripts 
+- `unit-test.yml` workflow to start the express server
+
 ## [v1.0.2] - 2024-12-21
 ### Added
 - Added markdown component